DE7405: Committing updates to the tagging of MCD so that the compression and size come up as floating point value not an int value.

+++ Discover-Bank-Phone/src/com/discover/mobile/bank/deposit/DepositSubmissionActivity.java
@@ -120,9 +120,9 @@ public class DepositSubmissionActivity extends BaseActivity implements Completio
 		bundle.putInt(BankTrackingHelper.TRACKING_IMAGE_WIDTH, frontImageWidth);
 
 		bundle.putString(BankTrackingHelper.TRACKING_IMAGE_SIZE, 
-				String.valueOf(frontImageCompressedSize/KB_CONVERSION));
+				String.valueOf((float)frontImageCompressedSize/KB_CONVERSION));
 		bundle.putString(BankTrackingHelper.TRACKING_IMAGE_COMPRESSION, 
-				String.valueOf(compression/PERCENTAGE_CONVERSION));
+				String.valueOf((float)compression/PERCENTAGE_CONVERSION));
 
 		bundle.putInt(BankTrackingHelper.TRACKING_IMAGE_AMOUNT, data.amount.value);
 		bundle.putInt(BankTrackingHelper.TRACKING_IMAGE_ACCOUNT, data.account);
